Add title screen
It's a little barebones but it'll do for now. The game will also go back to the title screen after the player dies.

The player's _just_shot_missile variable was set to true instead of false, so that it won't grab a click right when there's the transition between title.tscn and space.tscn.