Replace high score with leaderboard
When you beat your high score, a popup comes in asking for a name to put your high score by. This is stored in save data to keep a record of high scores.

To make the scene tree look a little bit cleaner, two more scenes were added that handle the entire leaderboard and separate records within the leaderboard, respectively.

If the save data is cleared within the options screen, the leaderboard will also update to reflect that.